For #9388, introduced SGCutReader.
Added a new SGCutReader class to read Cuts from SG from the code which was in the ShotGrid adapter.
Moved a couple of methods to utils to the reader.
